update test + refactor
