+++ src/pages/version/buildVariants/buildVariants.test.ts
@@ -1,7 +1,7 @@
 import { TaskStatus } from "types/task";
 import { groupTasksByUmbrellaStatus } from "./utils";
 
-describe("groupTasksByColor", () => {
+describe("groupTasksByUmbrellaStatus", () => {
   test("Seperates tasks into groups based on the color of the status", () => {
     const tasks = [
       {
@@ -23,24 +23,28 @@ describe("groupTasksByColor", () => {
         status: TaskStatus.Started,
       },
     ];
+
     expect(groupTasksByUmbrellaStatus(tasks)).toStrictEqual({
       "failed-umbrella": {
         count: 1,
         statuses: ["failed"],
         textColor: "#8F221B",
         fill: "#FCEBE2",
+        border: "#F9D3C5",
       },
       success: {
         count: 1,
         statuses: ["success"],
         textColor: "#116149",
         fill: "#E4F4E4",
+        border: "#C3E7CA",
       },
       "running-umbrella": {
         count: 1,
         statuses: ["started"],
         textColor: "#86681D",
         fill: "#FEF7E3",
+        border: "#FEF2C8",
       },
     });
   });
@@ -78,18 +82,21 @@ describe("groupTasksByColor", () => {
         statuses: ["failed", "task-timed-out"],
         textColor: "#8F221B",
         fill: "#FCEBE2",
+        border: "#F9D3C5",
       },
       success: {
         count: 1,
         statuses: ["success"],
         textColor: "#116149",
         fill: "#E4F4E4",
+        border: "#C3E7CA",
       },
       "running-umbrella": {
         count: 1,
         statuses: ["started"],
         textColor: "#86681D",
         fill: "#FEF7E3",
+        border: "#FEF2C8",
       },
     });
   });
@@ -127,18 +134,21 @@ describe("groupTasksByColor", () => {
         statuses: ["failed"],
         textColor: "#8F221B",
         fill: "#FCEBE2",
+        border: "#F9D3C5",
       },
       success: {
         count: 1,
         statuses: ["success"],
         textColor: "#116149",
         fill: "#E4F4E4",
+        border: "#C3E7CA",
       },
       "running-umbrella": {
         count: 1,
         statuses: ["started"],
         textColor: "#86681D",
         fill: "#FEF7E3",
+        border: "#FEF2C8",
       },
     });
   });
